function update, refactor comps
