+++ client/src/components/FlexScales/FlexTimeScale.jsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/core */
 import { css, jsx } from "@emotion/core";
-import { flexScale } from "../helpers/standardDefaults/flexTimeScale";
-import { Modal } from "./Modal";
+import { flexScale } from "../../helpers/standardDefaults/flexTimeScale";
+import { Modal } from "../General/Modal";
 
 const base = css`
   display: grid;
